garma: A package for estimating and foreasting Gegenbauer time series models.

The GARMA package provides the main function "garma" as well as print, summary, predict, forecast and plot/ggplot options.

The garma function is the main function for the garma package. Depending on the parameters it will calculate the parameter estimates for the GARMA process, and if available the standard errors (se's) for those parameters.

garma(
  x,
  order = list(0, 0, 0),
  k = 1,
  include.mean = (order[2] == 0),
  method = "Whittle",
  allow_neg_d = TRUE,
  maxeval = 10000,
  opt_method = "solnp",
  m_trunc = 50
)

Arguments

x
(num) This should be a numeric vector representing the process to estimate. A minimum length of 96 is required.

order
(list) This should be a list (similar to the stats::arima order parameter) which will give the order of the process to fit. The format should be list(p,d,q) where p, d, and q are all positive integers. p represents the degree of the autoregressive process to fit, q represents the order of the moving average process to fit and d is the (integer) differencing to apply prior to any fitting.

k
(int) This is the number of (multiplicative) Gegenbauer terms to fit. Only 0 or 1 are allowed in this version.

include.mean
(bool) A boolean value indicating whether a mean should be fit. Note if you have any differencing, then it generally does not make sense to fit a mean term. Because of this, the default here is to fit a mean time when d (in the "order" parmaeter) is zero and otherwise not.

method
(character) This defines the estimation method for the routine. The valid values are 'CSS', 'Whittle', 'QML' and 'WLL'. The default (Whittle) will generally return very accurate estimates quite quickly, provided the asumption of a Gaussian distribution is even approximately correct, and is probably the method of choice for most users. For the theory behind this, refer Giraitis et. al. (2001) 'CSS' is a conditional 'sum-of-squares' technique and can be quite slow. Reference: Chung (1996). 'QML' is a Quasi-Maximum-Likelihood technique, and can also be quite slow. Reference Dissanayake (2016). (k>1 is not supported for QML) 'WLL' is a new technique which appears to work well even if the \(\epsilon_{t}\) are highly skewed and/or have heavy tails (skewed and/or lepto-kurtic). However the asymptotic theory for the WLL method is not complete and so standard errors are not available for most parameters.

allow_neg_d
(bool) A boolean value indicating if a negative value is allowed for the fractional differencing component of the Gegenbauer term is allowed. This can be set to FALSE to force the routine to find a positive value.

maxeval
(int) the maximum function eveluations to be allowed during each optimisation.

opt_method
(character) This names the optimisation method used to find the parameter estimates. The default is to use 'solnp' from package Rsolnp, which has shown to have good performance in a wide range of circumstances. For some data or some models, however, other methods may work well. Supported algorithms include:

cobyla algorithm in package nloptr

directL algorithm in package nloptr

BBoptim from package BB

psoptim from package pso

hjkb from dfoptim package

nmkb from dfoptim package

solnp from Rsolnp package

best - this option evaluates all the above options in turn and picks the one which finds the lowest value of the objective. This can be quite time consuming to run, particularly for the 'CSS' method.

Note further that if you specify a k>1, then inequality constraints are required, and this will further limit the list of supported routines.

m_trunc
Used for the QML estimation method. This defines the AR-truncation point when evaluating the likelihood function. Refer to Dissanayake et. al. (2016) for details.

Value

An S3 object of class "garma_model".

References:
C Chung. A generalized fractionally integrated autoregressive moving-average process. Journal of Time Series Analysis, 17(2):111–140, 1996.
G Dissanayake, S Peiris, and T Proietti. State space modelling of Gegenbauer processes with long memory. Computational Statistics and Data Analysis, 100:115–130, 2016.
L Giraitis, J Hidalgo, and P Robinson. Gaussian estimation of parametric spectral density with unknown pole. The Annals of Statistics, 29(4):987–1023, 2001.

Details

The GARMA model is specified as $$\displaystyle{\phi(B)\prod_{i=1}^{k}(1-2u_{i}B+B^{2})^{d_{i}}(X_{t}-\mu)= \theta(B) \epsilon _{t}}$$

where

\(\phi(B)\) represents the short-memory Autoregressive component of order p,

\(\theta(B)\) represents the short-memory Moving Average component of order q,

\((1-2u_{i}B+B^{2})^{d_{i}}\) represents the long-memory Gegenbauer component (there may in general be k of these),

\(X_{t}\) represents the observed process,

\(\epsilon_{t}\) represents the random component of the model - these are assumed to be uncorrelated but identically distributed variates. Generally the routines in this package will work best if these have an approximate Gaussian distribution.

\(B\) represents the Backshift operator, defined by \(B X_{t}=X_{t-1}\).

when k=0, then this is just a short memory model as fit by the stats "arima" function.

Examples

data(AirPassengers)
ap  <- as.numeric(diff(AirPassengers,12))
print(garma(ap,order=c(9,1,0),k=0,method='CSS',include.mean=F))

#> 
#> Call:
#> garma(x = ap, order = c(9, 1, 0), k = 0, include.mean = F, method = "CSS")
#> 
#> Coefficients:
#>           ar1      ar2      ar3      ar4      ar5     ar6      ar7      ar8
#>       -0.3183  -0.0137  -0.1357  -0.2317  -0.0095  0.0338  -0.1480  -0.0896
#> s.e.   0.0860   0.0918   0.0916   0.0927   0.0962  0.0959   0.0938   0.0947
#>          ar9
#>       0.2224
#> s.e.  0.0979
#> 
#> 
#> sigma^2 estimated as 123.4608:  part log likelihood = -503.265174

# Compare with the built-in arima function
print(arima(ap,order=c(9,1,0),include.mean=F))

#> 
#> Call:
#> arima(x = ap, order = c(9, 1, 0), include.mean = F)
#> 
#> Coefficients:
#>           ar1      ar2      ar3      ar4      ar5     ar6      ar7      ar8
#>       -0.3157  -0.0131  -0.1354  -0.2238  -0.0083  0.0278  -0.1435  -0.0859
#> s.e.   0.0856   0.0915   0.0903   0.0914   0.0947  0.0939   0.0918   0.0924
#>          ar9
#>       0.2049
#> s.e.  0.0940
#> 
#> sigma^2 estimated as 123.3:  log likelihood = -501.71,  aic = 1023.42