The summary function provides a summary of a "garma_model" object.

The summary function provides a summary of a "garma_model" object.

# S3 method for garma_model
summary(mdl, verbose = TRUE)

Arguments

mdl
(garma_model) The garma_model from which to print the values.

verbose
(bool) whether to print out the verbose version of the model or not. Default: TRUE

Examples

data(AirPassengers)
ap  <- as.numeric(diff(AirPassengers,12))
mdl <- garma(ap,order=c(9,1,0),k=0,method='CSS',include.mean=F)
summary(mdl)

#> 
#> Call:
#> garma(x = ap, order = c(9, 1, 0), k = 0, include.mean = F, method = "CSS")
#> 
#> Summary of a Gegenbauer Time Series model.
#> 
#> Fit using CSS method.
#> Order=(9,1,0) k=0 
#> 
#> Optimisation.
#> Method:  solnp
#> Maxeval: 10000
#> Optimal Value found: 4.84556071
#> 
#> Coefficients:
#>           ar1      ar2      ar3      ar4      ar5     ar6      ar7      ar8
#>       -0.3183  -0.0137  -0.1357  -0.2317  -0.0095  0.0338  -0.1480  -0.0896
#> s.e.   0.0860   0.0918   0.0916   0.0927   0.0962  0.0959   0.0938   0.0947
#>          ar9
#>       0.2224
#> s.e.  0.0979
#> 
#> 
#> sigma^2 estimated as 123.4608:  part log likelihood = -503.265174